Phong upcode add CRUD cơ bản và đổi sang code first

+++ code/Garage-Finder-Backend/DataAccess/DTO/OrderDetailDTO.cs
@@ -8,7 +8,7 @@ namespace DataAccess.DTO
         public int OrderID { get; set; }
         public int ServiceID { get; set; }
         public string NameService { get; set; }
-        public float Cost { get; set; }
+        public double Cost { get; set; }
         public string Note { get; set; }
         public virtual ServiceDTO Service { get; set; }
         public int CategoryID { get; set; }
